% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_comparison.R
\name{benchmark_comparison}
\alias{benchmark_comparison}
\title{Benchmark Comparison}
\usage{
benchmark_comparison(success, total, benchmark)
}
\arguments{
\item{success}{success}

\item{total}{total}

\item{benchmark}{benchmark}
}
\value{
probability
}
\description{
Benchmark Comparison
}
\examples{
benchmark_comparison(10, 12, 0.7)
benchmark_comparison(60, 100, 0.7)
}
