% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_event.R
\name{benchmark_event}
\alias{benchmark_event}
\title{Compare Probability of an Event with Benchmark}
\usage{
benchmark_event(
  benchmark,
  event,
  total,
  event_type = "",
  notes = c("minimal", "technical", "none")
)
}
\arguments{
\item{benchmark}{benchmark}

\item{event}{event}

\item{total}{total}

\item{event_type}{Optional: a string describing the type of event. For example, success, failure, etc.}

\item{notes}{whether output should contain minimal or technical type of notes. Defaults to "minimal". Use "none" to turn off.}
}
\value{
list of event rate, probability, notes
}
\description{
Compare Probability of an Event with Benchmark
}
\examples{
benchmark_event(benchmark = 0.7,
                event = 10,
                total = 12,
                event_type = "success",
                notes = "minimal")
}
