% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_score.R
\name{benchmark_score}
\alias{benchmark_score}
\title{Compare Score with a Benchmark}
\usage{
benchmark_score(benchmark, scores, tail = "one", remove_missing = TRUE)
}
\arguments{
\item{benchmark}{benchmark}

\item{scores}{a column or vector of scores}

\item{tail}{one-tailed or two-tailed test}

\item{remove_missing}{TRUE/FALSE remove missing values? (default is TRUE)}
}
\value{
lower_ci, upper_ci, t, probability
}
\description{
Compare Score with a Benchmark
}
\examples{
scores <- 80 + 23 * scale(rnorm(172)) # 80 = mean, 23 = sd
benchmark_score(benchmark = 67, scores = scores)
}
