% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_time.R
\name{benchmark_time}
\alias{benchmark_time}
\title{Compare Time with a Benchmark}
\usage{
benchmark_time(benchmark, time, alpha, remove_missing = FALSE)
}
\arguments{
\item{benchmark}{benchmark}

\item{time}{a column or vector of time values}

\item{alpha}{alpha}

\item{remove_missing}{TRUE/FALSE remove missing values?}
}
\value{
lower_ci, upper_ci, t, probability
}
\description{
Compare Time with a Benchmark
}
\examples{
benchmark_time(time = c(60, 53, 70, 42, 62, 43, 81),
                       benchmark = 60,
                       alpha = 0.05)
}
