% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_means_between_groups.R
\name{compare_means_between_groups}
\alias{compare_means_between_groups}
\title{Compare Means Between Groups}
\usage{
compare_means_between_groups(
  data,
  variable,
  grouping_variable,
  groups,
  equal_variances = FALSE
)
}
\arguments{
\item{data}{data}

\item{variable}{variable}

\item{grouping_variable}{Group}

\item{groups}{Specify groups from grouping variable}

\item{equal_variances}{Specify whether variances are equal (TRUE/FALSE)}
}
\value{
results
}
\description{
Compare Means Between Groups
}
\examples{
compare_means_between_groups(mtcars, mpg, cyl, c(4, 6))
}
