% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_means_within_groups.R
\name{compare_means_within_groups}
\alias{compare_means_within_groups}
\title{Compare Means Within Groups}
\usage{
compare_means_within_groups(x, y, ...)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{...}{other arguments passed to paired t-test}
}
\value{
results
}
\description{
Compare Means Within Groups
}
