% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_rates_between_groups.R
\name{compare_rates_between_groups}
\alias{compare_rates_between_groups}
\title{Compare Rates Between Groups}
\usage{
compare_rates_between_groups(data, x, y, test)
}
\arguments{
\item{data}{data}

\item{x}{var 1}

\item{y}{var 2}

\item{test}{Type of test (fisher, n-1 two prop)}
}
\value{
results
}
\description{
Compare Rates Between Groups
}
