% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_rates_within_groups.R
\name{compare_rates_within_groups}
\alias{compare_rates_within_groups}
\title{Compare Rates Within Groups}
\usage{
compare_rates_within_groups(
  data,
  x,
  y,
  conf_level = 0.95,
  input,
  output = "console"
)
}
\arguments{
\item{data}{data}

\item{x}{var 1}

\item{y}{var 2}

\item{conf_level}{Confidence level}

\item{input}{input type currently only accepts "wide"}

\item{output}{Default is "console", also accepts "tibble"}
}
\value{
results
}
\description{
Compare Rates Within Groups
}
\examples{
A <- c(1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1)
B <- c(0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0)
data <- data.frame(A, B)
compare_rates_within_groups(data, A, B, input = "wide")
}
