% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_t.R
\name{dist_t}
\alias{dist_t}
\title{T Distribution}
\usage{
dist_t(t, df, tail)
}
\arguments{
\item{t}{t}

\item{df}{degrees of freedom}

\item{tail}{'one' or 'two'}
}
\value{
value
}
\description{
T Distribution
}
\examples{
dist_t(1.4, 2, "one")
dist_t(1.4, 2, "two")
}
