% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_concordant_discordant_pairs.R
\name{get_concordant_discordant_pairs}
\alias{get_concordant_discordant_pairs}
\title{Get concordant and discordant pairs for two variables}
\usage{
get_concordant_discordant_pairs(.data, id, var_1, var_2)
}
\arguments{
\item{.data}{= data}

\item{id}{= id column}

\item{var_1}{variable 1}

\item{var_2}{variable 2}
}
\value{
a data frame
}
\description{
Get concordant and discordant pairs for two variables
}
\examples{
mtcars$id <- seq.int(nrow(mtcars))
get_concordant_discordant_pairs(mtcars, id = id, var_1 = vs, var_2 = am)
}
