% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_confidence_intervals_event.R
\name{get_confidence_intervals_event}
\alias{get_confidence_intervals_event}
\title{Get Confidence Intervals for Event Rate}
\usage{
get_confidence_intervals_event(event, total, confidence_level)
}
\arguments{
\item{event}{event}

\item{total}{total}

\item{confidence_level}{confidence level as z value}
}
\value{
lower_ci, upper_ci
}
\description{
Get Confidence Intervals for Event Rate
}
\examples{
get_confidence_intervals_event(event = 80,
                               total = 100,
                               confidence_level = 1.96)
}
