% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_confidence_intervals_within_groups.R
\name{get_confidence_intervals_within_groups}
\alias{get_confidence_intervals_within_groups}
\title{Get Confidence Intervals Within Groups}
\usage{
get_confidence_intervals_within_groups(data, x, y, conf_level = 0.95)
}
\arguments{
\item{data}{data}

\item{x}{var 1}

\item{y}{var 2}

\item{conf_level}{Confidence level}
}
\value{
results
}
\description{
Get Confidence Intervals Within Groups
}
\examples{
A <- c(1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1)
B <- c(0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0)
data <- data.frame(A, B)
get_confidence_intervals_within_groups(data, A, B)
}
