% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_ci.R
\name{mean_ci}
\alias{mean_ci}
\title{Mean Confidence Intervals}
\usage{
mean_ci(x, alpha)
}
\arguments{
\item{x}{values}

\item{alpha}{alpha}
}
\value{
lower_ci, upper_ci
}
\description{
Mean Confidence Intervals
}
