% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_ci_2.R
\name{mean_ci_2}
\alias{mean_ci_2}
\title{Mean Confidence Intervals (Large Samples)}
\usage{
mean_ci_2(x, z)
}
\arguments{
\item{x}{values}

\item{z}{z value}
}
\value{
lower_ci, upper_ci
}
\description{
Mean Confidence Intervals (Large Samples)
}
