% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_t_test.R
\name{paired_t_test}
\alias{paired_t_test}
\title{Paired t-test}
\usage{
paired_t_test(x, y, ...)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{...}{other arguments passed to paired t-test}
}
\value{
results
}
\description{
Paired t-test
}
