% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sample_xy}
\alias{sample_xy}
\title{Sample XY - Example Dataset}
\format{
\subsection{\code{sample_xy}}{

A data frame with 100 rows and 2 columns:
\describe{
\item{x}{Column x}
\item{y}{Column y}
}
}
}
\usage{
sample_xy
}
\description{
Example data of two variables with uniform distribution
}
\keyword{datasets}
