% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_mean_ci_2.R
\name{stat_mean_ci_2}
\alias{stat_mean_ci_2}
\title{Mean Confidence Intervals (Large Samples)}
\usage{
stat_mean_ci_2(x, z)
}
\arguments{
\item{x}{values}

\item{z}{z value}
}
\value{
lower_ci, upper_ci
}
\description{
Mean Confidence Intervals (Large Samples)
}
\examples{
mean_ci_2(c(1, 2, 3, 4, 5, 6, 7), 1.96)
mean_ci_2(c(2, 4, 6, 8), 1.96)
}
