% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_observed_expected.R
\name{table_observed_expected}
\alias{table_observed_expected}
\title{Observed Expected Table}
\usage{
table_observed_expected(data, x, y)
}
\arguments{
\item{data}{data}

\item{x}{x}

\item{y}{y}
}
\value{
results
}
\description{
Observed Expected Table
}
