% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_completion.R
\name{task_completion}
\alias{task_completion}
\title{Task Completion}
\usage{
task_completion(success, failure = NULL, total, ci_z)
}
\arguments{
\item{success}{success}

\item{failure}{failure}

\item{total}{total}

\item{ci_z}{z value}
}
\value{
lower_ci, upper_ci
}
\description{
Task Completion
}
\examples{
task_completion(80, NULL, 100, 1.96)
task_completion(10, NULL, 12, 1.96)
}
