% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_time_benchmark.R
\name{task_time_benchmark}
\alias{task_time_benchmark}
\title{Task Time Benchmark}
\usage{
task_time_benchmark(x, benchmark, alpha)
}
\arguments{
\item{x}{a vector}

\item{benchmark}{benchmark}

\item{alpha}{alpha}
}
\value{
lower_ci, upper_ci, t, probability
}
\description{
Task Time Benchmark
}
\examples{
task_time_benchmark(c(1, 2, 3), 2, 0.05)
}
