% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fisher.R
\name{test_fisher}
\alias{test_fisher}
\title{Fisher's Test}
\usage{
test_fisher(data, x, y)
}
\arguments{
\item{data}{data}

\item{x}{x}

\item{y}{y}
}
\value{
results
}
\description{
Fisher's Test
}
\examples{
design = c("A","B")
complete = c(11, 5)
incomplete = c(1, 5)
data <- data.frame(design, complete, incomplete)
data <- data |> tidyr::pivot_longer(!design, names_to = "rate", values_to = "n") |>
  tidyr::uncount(n)
test_fisher(data, design, rate)
}
