% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mcnemar.R
\name{test_mcnemar}
\alias{test_mcnemar}
\title{McNemar Test}
\usage{
test_mcnemar(data, x, y)
}
\arguments{
\item{data}{data}

\item{x}{var 1}

\item{y}{var 2}

\item{id}{id or group}
}
\value{
results
}
\description{
McNemar Test
}
\examples{
A <- c(1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1)
B <- c(0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0)
data <- data.frame(A, B)
test_mcnemar(data, A, B)
}
