% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_n_1_two_prop.R
\name{test_n_1_two_prop}
\alias{test_n_1_two_prop}
\title{N-1 Two Proportions Test}
\usage{
test_n_1_two_prop(data, x, y, value, z)
}
\arguments{
\item{data}{data}

\item{x}{x}

\item{y}{y}

\item{value}{value}

\item{z}{z}
}
\value{
results
}
\description{
N-1 Two Proportions Test
}
