% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_t.R
\name{test_t}
\alias{test_t}
\title{T-test}
\usage{
test_t(x, y, ...)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{...}{other arguments passed to t-test}
}
\value{
results
}
\description{
T-test
}
\examples{

test_t(mtcars$mpg, mtcars$am)
}
