% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_t_paired.R
\name{test_t_paired}
\alias{test_t_paired}
\title{Paired t-test}
\usage{
test_t_paired(x, y, ...)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{...}{other arguments passed to paired t-test}
}
\value{
results
}
\description{
Paired t-test
}
