% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_wald_adj.R
\name{test_wald_adj}
\alias{test_wald_adj}
\title{Adjusted Wald Confidence Intervals}
\usage{
test_wald_adj(success, total, ci_z)
}
\arguments{
\item{success}{success}

\item{total}{total}

\item{ci_z}{z value}
}
\value{
lower_ci, upper_ci
}
\description{
Adjusted Wald Confidence Intervals
}
\examples{
test_wald_adj(10, 12, 1.96)
test_wald_adj(5, 7, 1.96)
}
