% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uxr-package.R
\docType{package}
\name{uxr-package}
\alias{uxr}
\alias{uxr-package}
\title{uxr: User Experience Research}
\description{
Provides convenience functions for user experience research with an emphasis on quantitative user experience testing and reporting. The functions are designed to translate statistical approaches to applied user experience research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://joe-chelladurai.github.io/uxr/}
  \item Report bugs at \url{https://github.com/joe-chelladurai/uxr/issues}
}

}
\author{
\strong{Maintainer}: Joe Chelladurai \email{joe.chelladurai@yahoo.com} (\href{https://orcid.org/0000-0001-8477-3753}{ORCID})

}
\keyword{internal}
