% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_wald_ci.R
\name{adj_wald_ci}
\alias{adj_wald_ci}
\title{Adjusted Wald Confidence Intervals}
\usage{
adj_wald_ci(success, total, ci_z)
}
\arguments{
\item{success}{success}

\item{total}{total}

\item{ci_z}{z value}
}
\value{
lower_ci, upper_ci
}
\description{
Adjusted Wald Confidence Intervals
}
\examples{
adj_wald_ci(10, 12, 1.96)
adj_wald_ci(5, 7, 1.96)
}
