% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_rates_within_groups.R
\name{compare_rates_within_groups}
\alias{compare_rates_within_groups}
\title{Compare Rates Within Groups}
\usage{
compare_rates_within_groups(data, x, y, z = 1.96)
}
\arguments{
\item{data}{data}

\item{x}{var 1}

\item{y}{var 2}

\item{z}{value}
}
\value{
results
}
\description{
Compare Rates Within Groups
}
\examples{
A <- c(1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1)
B <- c(0, 1, 1, 0, 0, 1, 1, 1, 0, 1, 0, 1, 0, 1, 0)
data <- data.frame(A, B)
}
