% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_concordant_discordant_pairs.R
\name{get_concordant_discordant_pairs}
\alias{get_concordant_discordant_pairs}
\title{Get concordant and discordant pairs for two variables}
\usage{
get_concordant_discordant_pairs(.data, id, x, y)
}
\arguments{
\item{.data}{= data}

\item{id}{= id column}

\item{x}{variable 1}

\item{y}{variable 2}
}
\value{
a data frame
}
\description{
Get concordant and discordant pairs for two variables
}
\examples{
mtcars$id <- seq.int(nrow(mtcars))
get_concordant_discordant_pairs(mtcars, id = id, x = vs, y = am)
}
