% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{wald_ci}
\alias{wald_ci}
\title{Wald Confidence Intervals}
\usage{
wald_ci(success, total, ci_z)
}
\arguments{
\item{success}{success}

\item{total}{total}

\item{ci_z}{z value}
}
\value{
lower_ci, upper_ci
}
\description{
Wald Confidence Intervals
}
\examples{
wald_ci(10, 12, 1.96)
wald_ci(5, 7, 1.96)
}
